requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
aiohttp==25.1.1
